requirements.txt
docopt
docutils
flask
psutil
kernelconfig
numpy
pyparted
peewee